fix(kafka-connect): parity block source offset serialisation

+++ apps/processing/connectors/sources/web3/src/main/kotlin/com/ethvm/kafka/connect/sources/web3/sources/ParityBlocksSource.kt
@@ -95,7 +95,7 @@ class ParityBlocksSource(
 
       val blockNumber = block.number
 
-      val partitionOffset = mapOf("blockNumber" to blockNumber)
+      val partitionOffset = mapOf("blockNumber" to blockNumber.toLong())
 
       val blockKeyRecord = CanonicalKeyRecord.newBuilder()
         .setNumberBI(blockNumber)
